[ThanhNT][11-TWELL-4-06-2024][DEVELOP]: update-for-all-service

+++ src/main/java/com/exe01/backend/service/impl/MajorServiceImpl.java
@@ -111,7 +111,7 @@ public class MajorServiceImpl implements IMajorService {
     }
 
     private int totalItemWithStatusActive() {
-        return (int) majorRepository.countByStatus(ConstStatus.ACTIVE_STATUS);
+        return majorRepository.countByStatus(ConstStatus.ACTIVE_STATUS);
     }
 
     @Override
